feature:
- 加入 MessageHeader

+++ src/TnyFramework.Net.DotNetty.AspNetCore/NetApplicationHostOptions.cs
@@ -9,7 +9,7 @@ namespace TnyFramework.Net.DotNetty.AspNetCore
 
         public string Name { get; set; }
 
-        public long ServerId { get; set; }
+        public int ServerId { get; set; }
 
         public string AppType { get; set; } = "default";
